refactor(filetransfer): split persistence responsibilities
